Rename BFF module to Gateway and update related configurations
