+++ apps/gateway/.gitattributes
