+++ apps/gateway/.gitignore
