+++ apps/gateway/build.gradle
@@ -6,7 +6,7 @@ plugins {
 
 group = 'org.shelfio'
 version = '0.0.1-SNAPSHOT'
-description = 'bff'
+description = 'gateway'
 
 java {
     toolchain {
